oa_module final/requirements.txt
# requirements.txt
langchain==0.1.9
langchain-openai==0.0.8
langchain-community==0.0.24
pydantic==2.6.1
streamlit==1.31.1
python-dotenv==1.0.1
typing-extensions==4.9.0
openai==1.12.0
pytest==8.0.1
markdown==3.5.2
jinja2==3.1.3
python-frontmatter==1.1.0